+++ jsolex/src/main/resources/whats-new.md
@@ -9,6 +9,7 @@ Here are the new features in this version:
 
 ## Changes since 2.4.0
 
+- Added "Doppler Eclipse" image
 - Improved resolution of reference spectral lines (1/100th of Angstrom)
 - Added ability to save GONG image by right-clicking on the image
 - Improved edge detection for improved images